fix: separate game logic
