+++ src/style.css
@@ -1,1 +1,3 @@
-
+body {
+  background-color: black;
+}
